Rename 'thumbnailer.go' -> 'thumbnail.go'
